Fixed multi-skin scale resetting to zero

+++ src/main/java/mchorse/mclib/utils/resources/FilteredResourceLocation.java
@@ -139,7 +139,7 @@ public class FilteredResourceLocation implements IWritableLocation
 
 		if (tag.hasKey("Scale"))
 		{
-			this.scale = tag.getInteger("Scale");
+			this.scale = tag.getFloat("Scale");
 		}
 
 		if (tag.hasKey("ScaleToLargest"))
